refactor: rename main files and improve project configuration
- Rename WindowsProject1.cpp/h to Client.cpp/h for better naming
- Update project name from TempProjectName to Client
- Upgrade to C++20 and C17 standards
- Disable precompiled headers for better compatibility
- Remove warning 4996 suppression
